pyproject.toml
[project]
name = "skysql-mcp"
version = "0.1.0"
description = "SkySQL MCP Server"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "fastmcp>=0.1.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
    "PyMySQL==1.1.0",
    "asyncio==3.4.3",
    "uvicorn>=0.27.0"
]
